Create custom field extension
Create a `md` field extension that render the markdown field from
`frontmatter` scope(in `md` file) into a html. To use this field
extension we must define types for GrapQL which field should use this
extension. For reference check `createSchemaCustomization` function in
the `gatsby-node.ts` file. Implemented based on the Gatsby docs [1].

Ref:
[1]: https://www.gatsbyjs.com/docs/reference/graphql-data-layer/schema-customization/#creating-custom-extensions