Add layer styles to the chakra theme
Create a `bodyMd` layer style based on the Threshold DS. Use this layer
style for a box where we inject the `html` via `dangerouslySetInnerHTML`
prop. We should't inject the `html` to `p`(the `BodyMd` component
renders the `p` tag) because the markdown widget in cms puts the text
inside the `p` tag- it's sematicly incorrect to render `p` tag inside
`p`.