Set up cms config for T proposals
The idea is an admin can add a proposal via cms- only url to a given
proposal is required and Gatsby will fetch the proposal data from
`forum.threshold.network` API under the hood. Optionally an admin can
override the proposal title and description in case of incorrect text
formatting or other issues. Advatages of this soulution:
- we can control which proposals are worth displaying on the T website,
- we wouldn't have fixed number of proposals to display. Eg. admin could
add more than 3 proposals and it will not require any updates in the
source code. It will depend on the admin.