Create a separate md file for tBTC partners
We want to display tBTC partners on main page and tBTC page. To do not
dupliacte list of partners in both `.md` files so here we store the
partners in a separate `.md` file. Here we also add a new field- `role`-
to specify which partner is `minter`, `guardina` or a `general` partner.