Fix error while updating the home page by admin
There is an error when we try to update the home page through admin panel:

```
Oops, you've missed a required field. Please complete before saving.
```

I happens, because the `posthogLabel` field was required for Liquidity Provider
Role section and it is not set by default in `stc/content/pages/index.md` file.
The posthog label for this button is not needed anyway so it was probably added
there by mistake.